Make GDL support a feature
