+++ Cargo.toml
@@ -17,7 +17,7 @@ byte-slice-cast = "1.1.0"
 delegate = "0.6.1"
 fast-float = "0.2.0"
 fxhash = "0.2.1"
-gdl = "0.2.4"
+gdl = { version = "0.2.4", optional = true }
 linereader = "0.4.0"
 log = "0.4.14"
 memmap2 = "0.5.0"
