+++ src/builder.rs
@@ -42,6 +42,7 @@ where
     _node: PhantomData<NI>,
 }
 
+#[doc(cfg(feature = "gdl"))]
 pub struct FromGdlString<NI>
 where
     NI: Idx,
@@ -51,6 +52,7 @@ where
     _node: PhantomData<NI>,
 }
 
+#[doc(cfg(feature = "gdl"))]
 pub struct FromGdlGraph<'a, NI>
 where
     NI: Idx,
@@ -268,6 +270,7 @@ impl GraphBuilder<Uninitialized> {
     /// assert_eq!(g.edge_count(), 2);
     /// assert_eq!(g.neighbors_with_values(0), &[Target::new(1, 0.42), Target::new(2, 13.37)]);
     /// ```
+    #[doc(cfg(feature = "gdl"))]
     pub fn gdl_str<NI, S>(self, gdl: S) -> GraphBuilder<FromGdlString<NI>>
     where
         NI: Idx,
@@ -295,7 +298,7 @@ impl GraphBuilder<Uninitialized> {
     /// ```
     /// use graph::prelude::*;
     ///
-    /// let gdl_graph = "(a)-->(),(a)-->()".parse::<gdl::Graph>().unwrap();
+    /// let gdl_graph = "(a)-->(),(a)-->()".parse::<::gdl::Graph>().unwrap();
     ///
     /// let g: DirectedCsrGraph<usize> = GraphBuilder::new()
     ///     .gdl_graph::<usize>(&gdl_graph)
@@ -309,7 +312,8 @@ impl GraphBuilder<Uninitialized> {
     ///
     /// assert_eq!(g.out_neighbors(id_a).len(), 2);
     /// ```
-    pub fn gdl_graph<NI>(self, gdl_graph: &gdl::Graph) -> GraphBuilder<FromGdlGraph<NI>>
+    #[doc(cfg(feature = "gdl"))]
+    pub fn gdl_graph<NI>(self, gdl_graph: &::gdl::Graph) -> GraphBuilder<FromGdlGraph<NI>>
     where
         NI: Idx,
     {
@@ -474,6 +478,7 @@ impl<NI: Idx, NV, EV> GraphBuilder<FromEdgeListAndNodeValues<NI, NV, EV>> {
     }
 }
 
+#[doc(cfg(feature = "gdl"))]
 impl<NI> GraphBuilder<FromGdlString<NI>>
 where
     NI: Idx,
@@ -489,6 +494,7 @@ where
     }
 }
 
+#[doc(cfg(feature = "gdl"))]
 impl<'a, NI> GraphBuilder<FromGdlGraph<'a, NI>>
 where
     NI: Idx,
